Add methods are general enough to be refactored into parent class DataAssistantResult
