+++ great_expectations/rule_based_profiler/types/data_assistant_result/data_assistant_result.py
@@ -1546,6 +1546,30 @@ class DataAssistantResult(SerializableDictDot):
         )
         return chart
 
+    def _create_display_chart_for_column_domain_expectation(
+        self,
+        expectation_configurations: List[ExpectationConfiguration],
+        attributed_metrics: Dict[Domain, Dict[str, ParameterNode]],
+        plot_mode: PlotMode,
+        sequential: bool,
+    ) -> List[alt.VConcatChart]:
+        column_dfs: List[pd.DataFrame] = self._create_column_dfs_for_charting(
+            attributed_metrics=attributed_metrics,
+            expectation_configurations=expectation_configurations,
+            plot_mode=plot_mode,
+        )
+
+        metric_type: alt.StandardType = AltairDataTypes.QUANTITATIVE.value
+        metric_name: str = "column_distinct_values_count"
+
+        return self._chart_column_values(
+            column_dfs=column_dfs,
+            metric_name=metric_name,
+            metric_type=metric_type,
+            plot_mode=plot_mode,
+            sequential=sequential,
+        )
+
     def _create_return_chart_for_column_domain_expectation(
         self,
         expectation_configuration: ExpectationConfiguration,
